+++ capycli/project/create_bom.py
@@ -9,6 +9,7 @@
 import logging
 import sys
 from typing import Any, Dict, List, Tuple
+import json
 
 from cyclonedx.model import ExternalReferenceType, HashAlgorithm
 from cyclonedx.model.bom import Bom
@@ -21,6 +22,7 @@ from capycli import get_logger
 from capycli.common.capycli_bom_support import CaPyCliBom, CycloneDxSupport, SbomCreator
 from capycli.common.print import print_red, print_text, print_yellow
 from capycli.common.purl_utils import PurlUtils
+from capycli.common.script_support import ScriptSupport
 from capycli.main.result_codes import ResultCode
 
 LOG = get_logger(__name__)
@@ -44,8 +46,9 @@ class CreateBom(capycli.common.script_base.ScriptBase):
                 return (key["mainlineState"], key["relation"])
         return ("", "")
 
-    def create_project_bom(self, project: Dict[str, Any]) -> List[Component]:
+    def create_project_bom(self, project: Dict[str, Any], create_controlfile: bool) -> Tuple[List, List]:
         bom: List[Component] = []
+        details: List[Dict] = []
         if not self.client:
             print_red("  No client!")
             sys.exit(ResultCode.RESULT_ERROR_ACCESSING_SW360)
@@ -55,6 +58,7 @@ class CreateBom(capycli.common.script_base.ScriptBase):
         for release in releases:
             print_text("   ", release["name"], release["version"])
             href = release["_links"]["self"]["href"]
+            sw360_id = self.client.get_id_from_href(href)
 
             try:
                 release_details = self.client.get_release_by_url(href)
@@ -97,15 +101,39 @@ class CreateBom(capycli.common.script_base.ScriptBase):
                 if "repository" in release_details and "url" in release_details["repository"]:
                     CycloneDxSupport.set_ext_ref(rel_item, ExternalReferenceType.VCS, comment="",
                                                  value=release_details["repository"]["url"])
+                attachments = self.get_release_attachments(release_details)
+                for attachment in attachments:
+                    at_type = attachment["attachmentType"]
+                    if at_type not in CaPyCliBom.FILE_COMMENTS:
+                        continue
+                    comment = CaPyCliBom.FILE_COMMENTS[at_type]
+                    at_data = self.client.get_attachment_by_url(attachment["_links"]["self"]["href"])
+                    if at_data.get("checkStatus") == "REJECTED":
+                        print_yellow("      WARNING: ignoring REJECTED attachment",
+                                     attachment["filename"])
+                        continue
+                    if at_type in ("SOURCE", "SOURCE_SELF", "BINARY", "BINARY_SELF"):
+                        ext_ref_type = ExternalReferenceType.DISTRIBUTION
+                    else:
+                        ext_ref_type = ExternalReferenceType.OTHER
+                        if create_controlfile:
+                            at_details = {
+                                "ComponentName": " ".join((release["name"], release["version"])),
+                                "Sw360Id": sw360_id,
+                                "Sw360AttachmentId": self.client.get_id_from_href(attachment["_links"]["self"]["href"])}
+                            for key in ("createdBy", "createdTeam", "createdOn", "createdComment", "checkStatus",
+                                        "checkedBy", "checkedTeam", "checkedOn", "checkedComment"):
+                                if key in at_data and at_data[key]:
+                                    at_details[key[0].upper() + key[1:]] = at_data[key]
 
-                for at_type, comment in (("SOURCE", CaPyCliBom.SOURCE_FILE_COMMENT),
-                                         ("BINARY", CaPyCliBom.BINARY_FILE_COMMENT)):
-                    attachments = self.get_release_attachments(release_details,
-                                                               (at_type, at_type + "_SELF"))  # type: ignore
-                    for attachment in attachments:
-                        CycloneDxSupport.set_ext_ref(rel_item, ExternalReferenceType.DISTRIBUTION,
-                                                     comment, attachment["filename"],
-                                                     HashAlgorithm.SHA_1, attachment.get("sha1", ""))
+                            if at_type == "COMPONENT_LICENSE_INFO_XML":
+                                at_details["CliFile"] = attachment["filename"]
+                            elif at_type == "CLEARING_REPORT":
+                                at_details["ReportFile"] = attachment["filename"]
+                            details.append(at_details)
+                    CycloneDxSupport.set_ext_ref(rel_item, ext_ref_type,
+                                                 comment, attachment["filename"],
+                                                 HashAlgorithm.SHA_1, attachment.get("sha1"))
 
             except SW360Error as swex:
                 print_red("    ERROR: unable to access project:" + repr(swex))
@@ -116,7 +144,6 @@ class CreateBom(capycli.common.script_base.ScriptBase):
                 CycloneDxSupport.set_property(rel_item, CycloneDxSupport.CDX_PROP_PROJ_STATE, mainline_state)
                 CycloneDxSupport.set_property(rel_item, CycloneDxSupport.CDX_PROP_PROJ_RELATION, relation)
 
-            sw360_id = self.client.get_id_from_href(href)
             CycloneDxSupport.set_property(rel_item, CycloneDxSupport.CDX_PROP_SW360ID, sw360_id)
 
             CycloneDxSupport.set_property(
@@ -128,9 +155,9 @@ class CreateBom(capycli.common.script_base.ScriptBase):
 
         # sub-projects are not handled at the moment
 
-        return bom
+        return bom, details
 
-    def create_project_cdx_bom(self, project_id: str) -> Bom:
+    def create_project_cdx_bom(self, project_id: str, create_controlfile: bool) -> Tuple[Bom, Dict]:
         if not self.client:
             print_red("  No client!")
             sys.exit(ResultCode.RESULT_ERROR_ACCESSING_SW360)
@@ -146,14 +173,19 @@ class CreateBom(capycli.common.script_base.ScriptBase):
 
         print_text("  Project name: " + project["name"] + ", " + project["version"])
 
-        cdx_components = self.create_project_bom(project)
+        cdx_components, control_components = self.create_project_bom(project, create_controlfile)
 
         creator = SbomCreator()
         sbom = creator.create(cdx_components, addlicense=True, addprofile=True, addtools=True,
                               name=project.get("name", ""), version=project.get("version", ""),
                               description=project.get("description", ""))
 
-        return sbom
+        controlfile = {
+            "ProjectName": ScriptSupport.get_full_name_from_dict(project, "name", "version"),
+            "Components": control_components
+        }
+
+        return sbom, controlfile
 
     def show_command_help(self) -> None:
         print("\nusage: CaPyCli project createbom [options]")
@@ -166,6 +198,7 @@ class CreateBom(capycli.common.script_base.ScriptBase):
   -name            name of the project, component or release
   -version         version of the project, component or release
   -o OUTPUTFILE    output file to write to
+  -ct CONTROLFILE  write control file for "bom DownloadAttachments" and "project CreateReadme"
         """)
 
         print()
@@ -212,7 +245,11 @@ class CreateBom(capycli.common.script_base.ScriptBase):
             sys.exit(ResultCode.RESULT_COMMAND_ERROR)
 
         if pid:
-            bom = self.create_project_cdx_bom(pid)
+            bom, controlfile = self.create_project_cdx_bom(pid, args.controlfile)
             CaPyCliBom.write_sbom(bom, args.outputfile)
+
+            if args.controlfile:
+                with open(args.controlfile, "w") as outfile:
+                    json.dump(controlfile, outfile, indent=2)
         else:
             print_yellow("  No matching project found")